Rename package virtualenv to virtualenv_rewrite
