+++ virtualenv_rewrite/__main__.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import, division, print_function
 
 from functools import partial
 
-from virtualenv.cli import cmd
+from virtualenv_rewrite.cli import cmd
 
 
 main = partial(cmd, prog_name="virtualenv")
