+++ virtualenv_rewrite/_compat.py
