+++ virtualenv_rewrite/_scripts/activate.bat
