+++ virtualenv_rewrite/_scripts/activate.csh
