+++ virtualenv_rewrite/_scripts/activate.fish
