+++ virtualenv_rewrite/_scripts/activate.ps1
