+++ virtualenv_rewrite/_scripts/activate.sh
