+++ virtualenv_rewrite/_scripts/activate_this.py
