+++ virtualenv_rewrite/_utils.py
