+++ virtualenv_rewrite/flavors/base.py
