+++ virtualenv_rewrite/flavors/posix.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import, division, print_function
 
 import posixpath
 
-from virtualenv.flavors.base import BaseFlavor
+from virtualenv_rewrite.flavors.base import BaseFlavor
 
 
 class PosixFlavor(BaseFlavor):
